// app/invoice-app/src/app/api-client/model/invoice.ts
/**
 * Invoice API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InvoiceItem } from './invoiceItem';
import { Customer } from './customer';


export interface Invoice { 
    id?: number;
    customer: Customer;
    description?: string;
    invoiceNumber: string;
    invoiceDate?: string;
    serviceDate?: string;
    paymentDate?: string;
    createDate: string;
    finalizedDate?: string;
    items?: Array<InvoiceItem>;
}